refactor(compliance): move control-mode classification into the evidence package
The package-layout gate freezes the set of top-level src/agent_bom/*.py
modules, and this branch introduced a new one. The module classifies what a
control's evidence means, which is what agent_bom.evidence already owns
alongside policy.py and scan_run.py, so it belongs there rather than as a
243rd top-level module.

Import-only change: no behaviour, no public API, no test expectations move.
