+++ src/agent_bom/cis_controls.py
@@ -9,7 +9,7 @@ safeguards — CIS-02.1 (software asset inventory), CIS-07.1
 scanning) — are IMPLEMENTED BY this scan: producing this SBOM and this scan is
 the evidence they operate. Tagging a finding onto them would fail the safeguard
 the finding proves is working, so they are scored from scan freshness instead
-(see :mod:`agent_bom.compliance_control_modes`).
+(see :mod:`agent_bom.evidence.control_modes`).
 
 NOTE: These are the generic CIS Controls v8 (cross-platform).  CIS also
 publishes platform-specific Benchmarks (e.g., CIS AWS Foundations, CIS GCP,
@@ -23,9 +23,9 @@ from __future__ import annotations
 
 from typing import TYPE_CHECKING
 
-from agent_bom.compliance_control_modes import finding_taggable_controls
 from agent_bom.constants import AI_PACKAGES as _AI_PACKAGES
 from agent_bom.constants import high_risk_severities
+from agent_bom.evidence.control_modes import finding_taggable_controls
 from agent_bom.risk_analyzer import ToolCapability, classify_mcp_tool
 
 if TYPE_CHECKING:
